chore: separate styling into own file
