+++ client/components/auth-form.js
@@ -13,48 +13,16 @@ import {
   Paper,
   Typography,
   withStyles
-} from './MUIComponents'
-
-const styles = theme => ({
-  main: {
-    width: 'auto',
-    display: 'block', // Fix IE 11 issue.
-    marginLeft: theme.spacing.unit * 3,
-    marginRight: theme.spacing.unit * 3,
-    [theme.breakpoints.up(400 + theme.spacing.unit * 3 * 2)]: {
-      width: 400,
-      marginLeft: 'auto',
-      marginRight: 'auto'
-    }
-  },
-  paper: {
-    marginTop: theme.spacing.unit * 8,
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    padding: `${theme.spacing.unit * 2}px ${theme.spacing.unit * 3}px ${theme
-      .spacing.unit * 3}px`
-  },
-  avatar: {
-    margin: theme.spacing.unit,
-    backgroundColor: theme.palette.secondary.main
-  },
-  form: {
-    width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing.unit
-  },
-  submit: {
-    marginTop: theme.spacing.unit * 3
-  }
-})
+} from '../util/MUIComponents'
+import styles from '../util/styling'
 
 const AuthForm = props => {
   const {name, displayName, handleSubmit, error, classes} = props
 
   return (
-    <main className={classes.main}>
+    <main className={classes.authMain}>
       <CssBaseline />
-      <Paper className={classes.paper}>
+      <Paper className={classes.authPaper}>
         <Avatar className={classes.avatar}>
           <LockOutlinedIcon />
         </Avatar>
