+++ client/components/user-home.js
@@ -4,27 +4,8 @@ import {connect} from 'react-redux'
 import InputOutputWrapper from './edit-page/input-output-wrapper'
 import editCode from './edit-code'
 import Carousel from './carousel'
-import {Typography, withStyles} from './MUIComponents'
-
-const styles = theme => ({
-  littleGrid: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justify: 'center',
-    spacing: 40
-  },
-  date: {
-    fontSize: '16px',
-    padding: '10px'
-  },
-  title: {
-    fontFamily: theme.typography.fontFamily[1],
-    fontSize: '40px',
-    color: '#789236',
-    padding: '10px'
-  }
-})
+import {Typography, withStyles} from '../util/MUIComponents'
+import styles from '../util/styling'
 
 /**
  * COMPONENT
